Struct cfile::CFile

pub struct CFile {
    pub owned: bool,
    // some fields omitted
}

A reference to an open stream on the filesystem.

An instance of a CFile can be read and/or written depending on what modes it was opened with. CFile also implement Seek to alter the logical cursor that the file contains internally.

Fields

owned

Methods

impl CFile

fn from_raw(f: *mut FILE, owned: bool) -> Result<CFile>

Constructs a CFile from a raw pointer.

fn open<P: AsRef<Path>>(path: P, mode: &str) -> Result<CFile>

opens the file whose name is the string pointed to by filename and associates a stream with it.

The argument mode points to a string beginning with one of the following sequences (Additional characters may follow these sequences.)

Modes

r Open text file for reading. The stream is positioned at the beginning of the file.

r+ Open for reading and writing. The stream is positioned at the beginning of the file.

w Truncate to zero length or create text file for writing. The stream is positioned at the beginning of the file.

w+ Open for reading and writing. The file is created if it does not exist, otherwise it is truncated. The stream is positioned at the beginning of the file.

a Open for writing. The file is created if it does not exist. The stream is positioned at the end of the file. Subsequent writes to the file will always end up at the then current end of file, irrespective of any intervening fseek(3) or similar.

a+ Open for reading and writing. The file is created if it does not exist. The stream is positioned at the end of the file. Subse- quent writes to the file will always end up at the then current end of file, irrespective of any intervening fseek(3) or similar.

The mode string can also include the letter b either as last charac- ter or as a character between the characters in any of the two-character strings described above. This is strictly for compatibility with ISO/IEC 9899:1990 (ISO C90) and has no effect; the b is ignored.

fn open_stream<S: AsRawFd>(s: &S, mode: &str) -> Result<CFile>

associates a stream with the existing file descriptor.

The mode of the stream must be compatible with the mode of the file descriptor.

fn open_stdin() -> Result<CFile>

open stdin as a read only stream

fn open_stdout() -> Result<CFile>

open stdout as a write only stream

fn open_stderr() -> Result<CFile>

open stderr as a write only stream

fn open_tmpfile() -> Result<CFile>

open a temporary file as a read/write stream

fn reopen(&self, filename: &str, mode: &str) -> Result<CFile>

opens the file whose name is the string pointed to by filename and associates the stream pointed to by stream with it.

The original stream (if it exists) is closed. The mode argument is used just as in the open() function.

fn stream(&self) -> *mut FILE

returns the raw pointer of the stream

fn position(&self) -> Result<u64>

returns the current position of the stream.

fn eof(&self) -> bool

tests the end-of-file indicator for the stream

fn errno(&self) -> i32

tests the error indicator for the stream

fn last_error(&self) -> Option<Error>

get the last error of the stream

fn clear_error(&self)

clears the end-of-file and error indicators for the stream

fn file_name(&self) -> Result<PathBuf>

returns the file name of the stream

fn metadata(&self) -> Result<Metadata>

Queries metadata about the underlying file.